requirements.txt
django
djangorestframework
 django-mongodb-backend
mysqlclient
 django-cors-headers
djangorestframework-simplejwt[crypto]